feat: JWT implementado, prubas unitarias superadas y endponts asegurados

+++ app/routers/themes.py
@@ -1,14 +1,23 @@
-from fastapi import APIRouter, HTTPException
+from fastapi import APIRouter, HTTPException, Depends
 from typing import List
 from app.schemas.theme import ThemeCreate, ThemeOut, ThemeUpdate
 from app.services import theme_service
+from app.core.security import get_current_user
+from app.models.user import User
 
 # Todas las rutas tendrán el prefijo "/themes" (ej: http://localhost:8000/themes/)
 # y aparecerán agrupadas bajo la etiqueta "Themes" en Swagger.
 router = APIRouter(prefix="/themes", tags=["Themes"])
 
 @router.post("/", response_model=ThemeOut)
-async def create_theme(theme: ThemeCreate):
+async def create_theme(
+        theme: ThemeCreate,
+        current_user: User = Depends(get_current_user)
+):
+    # Filtro de Administrador
+    if not current_user.is_admin:
+        raise HTTPException(status_code=403, detail="Privilegios insuficientes. Solo los administradores pueden crear temas.")
+
     # Recibe el esquema ThemeCreate (que solo exige el 'name')
     # y lo envía al servicio para guardarlo como un nuevo documento en MongoDB.
     return await theme_service.create_theme(theme)
@@ -23,7 +32,15 @@ async def get_all_themes():
     return await theme_service.get_all_themes()
 
 @router.put("/{theme_id}", response_model=ThemeOut)
-async def update_theme(theme_id: int, theme: ThemeUpdate):
+async def update_theme(
+        theme_id: int,
+        theme: ThemeUpdate,
+        current_user: User = Depends(get_current_user)
+):
+    # Filtro de Administrador
+    if not current_user.is_admin:
+        raise HTTPException(status_code=403, detail="Privilegios insuficientes. Solo los administradores pueden editar temas.")
+
     # Permite modificar un tema existente.
     # Busca por el 'theme_id' global.
     updated_theme = await theme_service.update_theme(theme_id, theme)
@@ -35,7 +52,13 @@ async def update_theme(theme_id: int, theme: ThemeUpdate):
     return updated_theme
 
 @router.delete("/{theme_id}")
-async def delete_theme(theme_id: int):
+async def delete_theme(
+        theme_id: int,
+        current_user: User = Depends(get_current_user)
+):
+    # Filtro de Administrador
+    if not current_user.is_admin:
+        raise HTTPException(status_code=403, detail="Privilegios insuficientes. Solo los administradores pueden borrar temas.")
     # Elimina un tema de la base de datos global.
     if not await theme_service.delete_theme(theme_id):
         raise HTTPException(status_code=404, detail="Tema no encontrado")
